Improve defaulting of max_num_emb; move basic helper function from feasibility to parallel_embeddings
